Code simplification and several little enhancements
