+++ custom_components/edata/manifest.json
@@ -7,6 +7,6 @@
   "documentation": "https://github.com/uvejota/homeassistant-edata/",
   "iot_class": "cloud_polling",
   "issue_tracker": "https://github.com/uvejota/homeassistant-edata/issues",
-  "requirements": ["e-data==1.2.6", "python-dateutil>=2.8.2"],
+  "requirements": ["e-data==1.2.7", "python-dateutil>=2.8.2"],
   "version": "2024.01.0"
 }
